conf/requirements.txt
amqp==5.2.0
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
binaryornot==0.4.4
celery==5.4.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cookiecutter==2.6.0
Django==5.0.7
django-admin-honeypot==1.1.0
django-axes==6.5.1
django-debug-toolbar==4.4.6
django-filter==24.3
django-jazzmin==3.0.0
django-render-partial==0.4
idna==3.7
Jinja2==3.1.4
kombu==5.4.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pika==1.3.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.1
redis==5.2.0
requests==2.32.3
rich==13.7.1
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
